requirements.txt
absl-py==0.8.1     
astor==0.8.0     
bleach==1.5.0     
cachetools==3.1.1     
certifi==2019.9.11 
chardet==3.0.4     
Click==7.0       
dask==2.8.0     
Flask==1.1.1     
Flask-Cors==3.0.8     
gast==0.3.2     
google-auth==1.7.1     
google-auth-oauthlib==0.4.1     
google-pasta==0.1.8     
grpcio==1.25.0    
gunicorn==20.0.4    
h5py==2.10.0    
html5lib==0.9999999 
idna==2.8       
itsdangerous==1.1.0     
Jinja2=2.10.3    
Keras-Applications==1.0.8     
Keras-Preprocessing==1.1.0     
Markdown=3.1.1     
MarkupSafe==1.1.1     
nltk==3.4.5     
numpy==1.17.4    
oauthlib==3.1.0     
opt-einsum==3.1.0     
paho-mqtt==1.5.0.dev0
Pillow==6.2.1     
pip==19.3.1    
protobuf==3.10.0    
pyasn1==0.4.8     
pyasn1-modules==0.2.7     
requests==2.22.0    
requests-oauthlib==1.3.0     
rsa==4.0       
selenium==3.141.0   
setuptools==41.6.0    
six==1.13.0    
tensorboard==2.0.1     
tensorflow==1.5.0     
tensorflow-estimator==2.0.1     
tensorflow-tensorboard==1.5.1     
termcolor==1.1.0     
tflearn==0.3.2     
urllib3==1.25.7    
Werkzeug==0.16.0    
wheel==0.33.6    
wrapt==1.11.2 